requirements.txt
python-dotenv
lyrics-extractor